Splitting EUtilities and EUtilities-Windows just for a better seperation.
