+++ include/EShortcutListener.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "EUtilities.hpp"
+#include "EUtilities-Windows.hpp"
 #include <functional>
 #include <vector>
 #include <thread>
